extract ApiHeader component, add version selector to SchemaList
